vton_scraper/requirements.txt
undetected-chromedriver>=3.5.5
selenium>=4.0.0
requests>=2.28.0
Pillow>=9.0.0
